11/1/18 commit

+++ app/models.py
@@ -4,7 +4,7 @@ import jsonfield
 
 class User(models.Model):
     name = models.CharField(max_length=20)
-    phone = models.PositiveIntegerField(null=True)
+    phone = models.BigIntegerField(null=True)
     address = models.TextField()
     password = models.CharField(max_length=10)
     cart = jsonfield.JSONField(default={})
